+++ v3/crates/metadata-resolve/src/helpers/boolean_expression.rs
@@ -18,6 +18,7 @@ use std::collections::BTreeMap;
 // we want to know
 // a) does each scalar have mappings for our data connector?
 // b) if we used nested objects, does the data connector have the correct capability?
+// c) if we used nested arrays, does the data connector have the correct capability?
 pub(crate) fn validate_data_connector_with_object_boolean_expression_type(
     data_connector: &data_connectors::DataConnectorLink,
     source_type_mappings: &BTreeMap<Qualified<CustomTypeName>, object_types::TypeMapping>,
@@ -25,7 +26,10 @@ pub(crate) fn validate_data_connector_with_object_boolean_expression_type(
     boolean_expression_types: &boolean_expressions::BooleanExpressionTypes,
     object_types: &BTreeMap<Qualified<CustomTypeName>, relationships::ObjectTypeWithRelationships>,
     models: &IndexMap<Qualified<ModelName>, models::Model>,
-) -> Result<(), Error> {
+) -> Result<Vec<boolean_expressions::BooleanExpressionIssue>, Error> {
+    // collect any issues found whilst resolving
+    let mut issues = vec![];
+
     if let Some(graphql_config) = &object_boolean_expression_type.graphql {
         for (field_name, object_comparison_expression_info) in &graphql_config.object_fields {
             // look up the leaf boolean expression type
@@ -41,9 +45,12 @@ pub(crate) fn validate_data_connector_with_object_boolean_expression_type(
                     })
                 })?;
 
-            // this must be a nested object, so let's check our data connector is ready for that
-            if !data_connector.capabilities.supports_nested_object_filtering {
-                return Err(
+            match object_comparison_expression_info.field_kind {
+                boolean_expressions::ObjectComparisonKind::Object => {
+                    // throw an error if our data connector does not support filtering nested
+                    // objects
+                    if !data_connector.capabilities.supports_nested_object_filtering {
+                        return Err(
                     Error::from(boolean_expressions::BooleanExpressionError::NoNestedObjectFilteringCapabilitiesDefined {
                         parent_type_name: object_boolean_expression_type.name.clone(),
                         nested_type_name: object_comparison_expression_info
@@ -51,6 +58,21 @@ pub(crate) fn validate_data_connector_with_object_boolean_expression_type(
                             .clone(),
                         data_connector_name: data_connector.name.clone(),
                         }));
+                    }
+                }
+                boolean_expressions::ObjectComparisonKind::Array => {
+                    // raise a warning if our data connector does not support filtering nested arrays
+                    if !data_connector.capabilities.supports_nested_array_filtering {
+                        issues.push(
+                    boolean_expressions::BooleanExpressionIssue::NoNestedArrayFilteringCapabilitiesDefined {
+                        parent_type_name: object_boolean_expression_type.name.clone(),
+                        nested_type_name: object_comparison_expression_info
+                            .object_type_name
+                            .clone(),
+                        data_connector_name: data_connector.name.clone(),
+                        });
+                    }
+                }
             }
 
             // This nested object field should not contain any relationship comparisons
@@ -69,14 +91,14 @@ pub(crate) fn validate_data_connector_with_object_boolean_expression_type(
             }
 
             // continue checking the nested object...
-            validate_data_connector_with_object_boolean_expression_type(
+            issues.extend(validate_data_connector_with_object_boolean_expression_type(
                 data_connector,
                 source_type_mappings,
                 leaf_boolean_expression,
                 boolean_expression_types,
                 object_types,
                 models,
-            )?;
+            )?);
         }
 
         for (field_name, comparison_expression_info) in &graphql_config.scalar_fields {
@@ -116,7 +138,7 @@ pub(crate) fn validate_data_connector_with_object_boolean_expression_type(
         }
     }
 
-    Ok(())
+    Ok(issues)
 }
 
 // validate comparable relationship field against data connector