+++ v3/crates/metadata-resolve/src/stages/boolean_expressions/types.rs
@@ -1,4 +1,5 @@
 use crate::stages::scalar_boolean_expressions;
+use crate::types::error::ShouldBeAnError;
 use crate::types::subgraph::{Qualified, QualifiedTypeReference};
 use lang_graphql::ast::common as ast;
 use open_dds::{
@@ -10,8 +11,24 @@ use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
 
 #[derive(Debug, thiserror::Error, Serialize, Deserialize, Clone, PartialEq, Eq)]
-pub enum BooleanExpressionIssue {}
+pub enum BooleanExpressionIssue {
+    #[error("The data connector {data_connector_name} cannot be used for filtering nested array {nested_type_name:} within {parent_type_name:} as it has not defined any capabilities for nested array filtering")]
+    NoNestedArrayFilteringCapabilitiesDefined {
+        parent_type_name: Qualified<CustomTypeName>,
+        nested_type_name: Qualified<CustomTypeName>,
+        data_connector_name: Qualified<DataConnectorName>,
+    },
+}
 
+impl ShouldBeAnError for BooleanExpressionIssue {
+    fn should_be_an_error(&self, flags: &open_dds::flags::Flags) -> bool {
+        match self {
+            BooleanExpressionIssue::NoNestedArrayFilteringCapabilitiesDefined { .. } => {
+                flags.require_nested_array_filtering_capability
+            }
+        }
+    }
+}
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub struct BooleanExpressionTypes {
     pub objects: BTreeMap<Qualified<CustomTypeName>, ResolvedObjectBooleanExpressionType>,