+++ v3/crates/metadata-resolve/src/stages/models_graphql/filter.rs
@@ -22,7 +22,13 @@ pub(crate) fn resolve_filter_expression_type(
     boolean_expression_types: &boolean_expressions::BooleanExpressionTypes,
     object_types: &BTreeMap<Qualified<CustomTypeName>, relationships::ObjectTypeWithRelationships>,
     models: &IndexMap<Qualified<ModelName>, models::Model>,
-) -> Result<ModelExpressionType, Error> {
+) -> Result<
+    (
+        ModelExpressionType,
+        Vec<boolean_expressions::BooleanExpressionIssue>,
+    ),
+    Error,
+> {
     match object_boolean_expression_types.get(boolean_expression_type_name) {
         Some(object_boolean_expression_type) => {
             // we're using an old ObjectBooleanExpressionType kind
@@ -67,8 +73,11 @@ pub(crate) fn resolve_filter_expression_type(
                 });
             }
 
-            Ok(ModelExpressionType::ObjectBooleanExpressionType(
-                object_boolean_expression_type.clone(),
+            Ok((
+                ModelExpressionType::ObjectBooleanExpressionType(
+                    object_boolean_expression_type.clone(),
+                ),
+                vec![],
             ))
         }
         None => {
@@ -79,7 +88,7 @@ pub(crate) fn resolve_filter_expression_type(
             {
                 Some(boolean_expression_object_type) => {
                     // we're using the new style of BooleanExpressionType
-                    validate_data_connector_with_object_boolean_expression_type(
+                    let data_connector_issues = validate_data_connector_with_object_boolean_expression_type(
                         &model_source.data_connector,
                         &model_source.type_mappings,
                         boolean_expression_object_type,
@@ -87,10 +96,9 @@ pub(crate) fn resolve_filter_expression_type(
                         object_types,
                         models,
                     )?;
-
-                    Ok(ModelExpressionType::BooleanExpressionType(
+                    Ok((ModelExpressionType::BooleanExpressionType(
                         boolean_expression_object_type.clone(),
-                    ))
+                    ),data_connector_issues))
                 }
                 None => Err(Error::from(
                     boolean_expressions::BooleanExpressionError::UnknownBooleanExpressionTypeInModel {